Vagrant Automatic Dev Setup
Due to the number of dependencies, Oracle driver requirements, and
general Ruby/RVM foolishness, setting up a dev environment was
difficult. This PR introduces a Vagrant configuration file that
automatically sets-up a dev environment, ready to be worked on.

Additionally, it gives us a canonical and quick way to make
Wrektranet2 platforms, if we ever need to redeploy from scratch.